+++ domain/src/main/java/com/example/domain/model/PublicationModel.kt
@@ -2,7 +2,7 @@ package com.example.domain.model
 
 data class PublicationModel(
     val id: Int,
-    val channelId: Int,
+    val channelId: Long,
     val text: String,
     val comment: String,
     val scheduledAt: String?,
